requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
rank-bm25>=0.2.2
openai>=1.10.0
streamlit>=1.28.0
plotly>=5.17.0
pyyaml>=6.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
